source/zpixfont.cpp font.h : correction for >16 bit unicode

+++ source/font.h
@@ -3,7 +3,7 @@
  */
 #include <stdint.h>
 
-extern const uint16_t charcode[];
+extern const uint32_t charcode[];
 extern const uint8_t glyph[][18];
 
 void ExpandGlyph(const uint8_t glyph[18], uint16_t *bitmap);
